requirements.txt
openai>=1.0.0
requests
selenium
python-dotenv
beautifulsoup4
pandas
Pillow
flask
flask-cors
openpyxl
pymongo
langchain
langchain-openai
langchain-community
sentry-sdk[flask]>=1.40.0